LPS-193648 refactor CSDiagramSettingImageUploadResponseHandler.java
move and create inner class in UploadCSDiagramSettingImageMVCActionCommand.java and delete handler class
